requirements.txt
 reportlab 
 django-cors-headers
 django_extensions 
 djangorestframework 
 django-filter  